test-requirements.txt
oslotest>=3.2.0 # Apache-2.0
PyYAML>=6.0.0 # MIT
stestr>=2.0.0 # Apache-2.0
